+++ src/RunOperators/CaseEditor.xml
@@ -4,9 +4,8 @@
 
         <test name="myTest">
             <parameter name="elements" value="Notebook"> </parameter >
+            <parameter name="msg" value="Adobe Photoshop CS4">  </parameter>
             <classes>
-
-
                 <class name="RunOperators.TestCase1"> </class>
                 <class name="RunOperators.TestCase2"> </class>
                 <class name="RunOperators.TestCase3"> </class>
@@ -14,23 +13,7 @@
                 <class name="RunOperators.TestCase4_2"> </class>
                 <class name="RunOperators.TestCase5"> </class>
                 <class name="RunOperators.TestCase6"> </class>
+                <class name="RunOperators.TestCase7"> </class>
             </classes>
         </test>
-
-
-        <test name="myTest1">
-
-                <parameter name="msg" value="Adobe Photoshop CS4">  </parameter>
-            <classes>
-
-                    <class name="RunOperators.TestCase7">
-                        <methods>
-                            <include name="US_07_Search">  </include>
-                        </methods>
-                    </class>
-
-        </classes>
-
-
-    </test>
 </suite>
