+++ src/RunOperators/TestCase4_2.java
@@ -2,6 +2,12 @@ package RunOperators;
 
 import Elements.PrElement;
 import Utility.BaseDriverMethod;
+import org.openqa.selenium.Keys;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.testng.Assert;
+import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 
 import java.time.Duration;
@@ -11,9 +17,8 @@ import java.util.List;
 public class TestCase4_2 extends BaseDriverMethod {
     @Test
     @Parameters("elements")
-    public void US_04_TabMenu2(String element){
+    public void US_04_TabMenu2(String elements){
         PrElement pr=new PrElement();
-        WebDriverWait wait=new WebDriverWait(driver, Duration.ofSeconds(20));
 
         for (int i = 0; i < pr.listheader.size(); i++) {
             WebElement tabMenu = pr.listheader.get(i);
@@ -26,12 +31,12 @@ public class TestCase4_2 extends BaseDriverMethod {
 
 
 
-        pr.search.sendKeys(element, Keys.ENTER);
+        pr.search.sendKeys(elements, Keys.ENTER);
 
         // random bir ürün seçildi .
 
-        int randomSelect = NoopElement.randomGenerator(pr.productTitle.size());
-        wait.until(ExpectedConditions.urlContains(element));
+        int randomSelect = PrElement.randomGenerator(pr.productTitle.size());
+        wait.until(ExpectedConditions.urlContains(elements));
 
         List<String> stringList = new ArrayList<>();
         for (WebElement element : pr.productTitle) {
